+++ bindings/python/src/tokenizer.rs
@@ -629,7 +629,7 @@ impl PyTokenizer {
     ///     :obj:`Dict[str, int]`: The vocabulary
     #[args(with_added_tokens = true)]
     #[text_signature = "(self, with_added_tokens=True)"]
-    fn get_vocab(&self, with_added_tokens: bool) -> HashMap<String, u32> {
+    fn get_vocab(&self, with_added_tokens: bool) -> HashMap<String, u64> {
         self.tokenizer.get_vocab(with_added_tokens)
     }
 
@@ -1000,7 +1000,7 @@ impl PyTokenizer {
     ///     :obj:`str`: The decoded string
     #[args(skip_special_tokens = true)]
     #[text_signature = "(self, ids, skip_special_tokens=True)"]
-    fn decode(&self, ids: Vec<u32>, skip_special_tokens: bool) -> PyResult<String> {
+    fn decode(&self, ids: Vec<u64>, skip_special_tokens: bool) -> PyResult<String> {
         ToPyResult(self.tokenizer.decode(ids, skip_special_tokens)).into()
     }
 
@@ -1019,7 +1019,7 @@ impl PyTokenizer {
     #[text_signature = "(self, sequences, skip_special_tokens=True)"]
     fn decode_batch(
         &self,
-        sequences: Vec<Vec<u32>>,
+        sequences: Vec<Vec<u64>>,
         skip_special_tokens: bool,
     ) -> PyResult<Vec<String>> {
         let gil = Python::acquire_gil();
@@ -1037,7 +1037,7 @@ impl PyTokenizer {
     /// Returns:
     ///     :obj:`Optional[int]`: An optional id, :obj:`None` if out of vocabulary
     #[text_signature = "(self, token)"]
-    fn token_to_id(&self, token: &str) -> Option<u32> {
+    fn token_to_id(&self, token: &str) -> Option<u64> {
         self.tokenizer.token_to_id(token)
     }
 
@@ -1050,7 +1050,7 @@ impl PyTokenizer {
     /// Returns:
     ///     :obj:`Optional[str]`: An optional token, :obj:`None` if out of vocabulary
     #[text_signature = "(self, id)"]
-    fn id_to_token(&self, id: u32) -> Option<String> {
+    fn id_to_token(&self, id: u64) -> Option<String> {
         self.tokenizer.id_to_token(id)
     }
